+++ cookbooks/office/recipes/masterpdf.rb
@@ -20,7 +20,6 @@
 
 
 office = node[:apps][:office]
-box = node[:box]
 
 # Complete solution for view, print and edit PDF and XPS files
 install_app "masterpdf" do
@@ -30,12 +29,12 @@ end
 launcher "masterpdf" do
   template "/masterpdf/masterpdf.desktop.erb"
   variables(
-    :exec => "sh -c '~/#{box['folders']['apps']}/masterpdf/pdfeditor'"
+    :exec => "sh -c '~/#{node[:box][:folders][:apps]}/masterpdf/pdfeditor'"
   )
 end
 
 uninstaller "masterpdf" do
-  template "/masterpdf/uninstall_masterpdf-#{box['lang']}.sh.erb"
+  template "/masterpdf/uninstall_masterpdf-#{node[:box][:lang]}.sh.erb"
   variables(
     :app     => "masterpdf",
     :website => office['profiles']['masterpdf']['website']